refactor BenchmarkBase
