+++ sql/catalyst/src/test/scala/org/apache/spark/sql/HashBenchmark.scala
@@ -20,8 +20,8 @@ package org.apache.spark.sql
 import org.apache.spark.sql.catalyst.encoders.RowEncoder
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.expressions.codegen.GenerateSafeProjection
+import org.apache.spark.sql.execution.benchmark.Benchmark
 import org.apache.spark.sql.types._
-import org.apache.spark.util.Benchmark
 
 /**
  * Benchmark for the previous interpreted hash function(InternalRow.hashCode) vs codegened
