+++ sql/catalyst/src/test/scala/org/apache/spark/sql/HashByteArrayBenchmark.scala
@@ -20,9 +20,9 @@ package org.apache.spark.sql
 import java.util.Random
 
 import org.apache.spark.sql.catalyst.expressions.{HiveHasher, XXH64}
+import org.apache.spark.sql.execution.benchmark.Benchmark
 import org.apache.spark.unsafe.Platform
 import org.apache.spark.unsafe.hash.Murmur3_x86_32
-import org.apache.spark.util.Benchmark
 
 /**
  * Synthetic benchmark for MurMurHash 3 and xxHash64.
